feat(conditionally-query-cell): Added cache clearing when the component unmounts. Also separate thunk concerns to default query execution when files are undefined
